+++ libs/langchain-core/src/testing/index.ts
@@ -1,0 +1,2 @@
+export * from "./matchers.js";
+export { fakeModel } from "./fake_model_builder.js";