+++ libs/providers/langchain-google-gauth/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @langchain/google-gauth
 
+## 2.1.24
+
+### Patch Changes
+
+- Updated dependencies [[`745b09a`](https://github.com/langchain-ai/langchainjs/commit/745b09a1b0c92643cbc99e773012924948438629)]:
+  - @langchain/google-common@2.1.24
+
+## 2.1.23
+
+### Patch Changes
+
+- Updated dependencies [[`d43194b`](https://github.com/langchain-ai/langchainjs/commit/d43194b625f19acf79cbe958300dadb7b0a6a89d)]:
+  - @langchain/google-common@2.1.23
+
 ## 2.1.22
 
 ### Patch Changes